requirements.txt
flask
numpy
pillow
gunicorn
joblib
tensorflow-cpu
keras
